this idea of the enum for geom + the commited scene rep works
There's a bit of awkwardness left to fix but this has got me just about
exactly where I want the code to work. The commit returning a "committed
scene" which then blocks out mutation of the scene b/c the committed
scene borrows it works great. The scene owning the geom also works,
since the Embree doc mentions a geom can only be in one scene at a time
anyway.